adjust query so recurring events display in their expected order, add some JS [touch: 1945]
make an adjustment to the query so by default events will be ordered by the recurring id, then the date. the only thing I'm seeing that may be an issue is non-recurring events will display in the list before recurring events. Also added a bit of jQuery that makes it so you don't have to add any CSS to hide the dates.